scripts/requirements.txt
pypdn>=1.0.6
Pillow>=9.2.0
numpy>=1.23.1
